refactor: remplacer la gestion du filterset par ngrx refactor: remplacer les bouteilles filtrées par ngrx refactor: séparer filterset de la distribution
WIP
