+++ src/components/search-bar/search-bar.ts
@@ -1,5 +1,5 @@
 import {Component, EventEmitter, Input, OnChanges, Output, SimpleChanges} from '@angular/core';
-import {FilterSet} from '../distribution/distribution';
+import {FilterSet} from '../distribution/filterset';
 
 /**
  * Generated class for the SearchBarComponent component.
